Winged Steed of the Ebon Blade northrend fix
Winged Steed of the Ebon Blade gets incorrectly categorised as a ground mount, causes issues and bots will not be able to ground mount if they have this mount in the account spellbook.

There may by other scaling mounts that have the same issue, either we blacklist them all or figure out a way to correctly read them as flyers.