New update
Summary of changes implemented in the code:

Personal cooldown of 5 seconds to call or log out bots.

Cooldown is only reset if the bot is actually called/logged out.

Delay of 5 seconds for the bot to fully initialize in the world.

Same logic applied to logout: does not log out the bot if it is in cooldown.

Cooldown is stored by master (masterAccountId).