Update playerbots.conf.dist
I'm doing this because the list that is currently implemented is, at best, incomplete. There are often multiple IDs for the same object name, and all of them need to be included to effectively block access to all those objects. So I extracted all lootable game objects (type 3) from the world database and used ChatGPT to create a list from them.
The IDs are grouped in the commented section for easier readability.

If possible, please double-check that the IDs are correct.
I'm already testing it on my own server and it seems to work perfectly. I've been playing for three hours now, and the bots haven't stolen anything from me at all. But please check it yourself as well.

If anything needs to be removed, please let me know. I'll take a look at it.

I would also like to suggest that this list be integrated into the Playerbots database; that way, the configuration file wouldn't become so bloated.